auth consumer return nul

+++ back/cliente/src/main/java/com/bantads/cliente/cliente/repositories/ClienteRepository.java
@@ -10,6 +10,6 @@ public interface ClienteRepository extends JpaRepository<Cliente, Long> {
 
   // Cliente findByEmail(String email);
 
-  List<Cliente> findByGerente(Long gerente);
+  List<Cliente> findByGerente(int gerente);
 
 }
